+++ ocdskit/exceptions.py
@@ -23,6 +23,10 @@ class UnknownFormatError(OCDSKitError):
     """Raised if the format of a file can't be determined"""
 
 
+class UnknownVersionError(OCDSKitError):
+    """Raised if the OCDS version is not recognized"""
+
+
 class MissingOcidKeyError(OCDSKitError, KeyError):
     """Raised if a release to be merged is missing an ``ocid`` field"""